+++ constants/triggers.js
@@ -96,7 +96,7 @@ export const BAYOU_SLUDGE_MESSAGE = `${RESET}${GREEN}A swampy mass of slime emer
 export const ALLIGATOR_MESSAGE = `${RESET}${GREEN}A long snout breaks the surface of the water. It's an Alligator!${RESET}`;
 export const TITANOBOA_MESSAGE = `${RESET}${GREEN}${RESET}${RED}${BOLD}A massive Titanoboa surfaces. It's body stretches as far as the eye can see.${RESET}`; // &r&a&r&c&lA massive Titanoboa surfaces. It's body stretches as far as the eye can see.&r
 export const BLUE_RINGED_OCTOPUS_MESSAGE = `${RESET}${GREEN}A garish set of tentacles arise. It's a Blue Ringed Octopus!${RESET}`;
-export const WIKI_TIKI_MESSAGE = `${RESET}${RED}${RESET}${RED}${BOLD}The water bubbles and froths. A massive form emerges- you have disturbed the Wiki Tiki! You shall pay the price.${RESET}`;
+export const WIKI_TIKI_MESSAGE = `${RESET}${RED}${RESET}${RED}${BOLD}The water bubbles and froths. A massive form emerges- you have disturbed the Wiki Tiki! You shall pay the price.${RESET}`; // &r&c&r&c&lThe water bubbles and froths. A massive form emerges- you have disturbed the Wiki Tiki! You shall pay the price.&r
 
 // DROPS
 
@@ -948,7 +948,7 @@ export const KILLED_BY_TRIGGERS = [
     },
 ];
 
-// All mobs that can spawn on Crimson Isle except Thunder / Lord Jawbus
+// All sea creatures that can spawn on Crimson Isle except creatures tracked by Crimson Isle tracker
 export const REGULAR_CRIMSON_CATCH_TRIGGERS = [
     {
         trigger: MAGMA_SLUG_MESSAGE
@@ -972,19 +972,10 @@ export const REGULAR_CRIMSON_CATCH_TRIGGERS = [
         trigger: TAURUS_MESSAGE
     },
     {
-        trigger: PLHLEGBLAST_MESSAGE
-    },
-    {
-        trigger: RAGNAROK_MESSAGE
-    },
-    {
         trigger: FRIED_CHICKEN_MESSAGE
     },
     {
         trigger: FIREPROOF_WITCH_MESSAGE
-    },
-    {
-        trigger: FIERY_SCUTTLER_MESSAGE
     },
 ];
 
@@ -1070,6 +1061,103 @@ export const REGULAR_JERRY_WORKSHOP_CATCH_TRIGGERS = [
     },
     {
         trigger: GRIM_REAPER_MESSAGE
+    },
+];
+
+// All sea creatures that can spawn in Water hotspot / Bayou except Wiki Tiki / Titanoboa
+export const REGULAR_WATER_HOTSPOT_AND_BAYOU_CATCH_TRIGGERS = [
+    {
+        trigger: SQUID_MESSAGE
+    },
+    {
+        trigger: SEA_WALKER_MESSAGE
+    },
+    {
+        trigger: SEA_GUARDIAN_MESSAGE
+    },
+    {
+        trigger: SEA_WITCH_MESSAGE
+    },
+    {
+        trigger: SEA_ARCHER_MESSAGE
+    },
+    {
+        trigger: RIDER_OF_THE_DEEP_MESSAGE
+    },
+    {
+        trigger: CATFISH_MESSAGE
+    },
+    {
+        trigger: SEA_LEECH_MESSAGE
+    },
+    {
+        trigger: GUARDIAN_DEFENDER_MESSAGE
+    },
+    {
+        trigger: DEEP_SEA_PROTECTOR_MESSAGE
+    },
+    {
+        trigger: WATER_HYDRA_MESSAGE
+    },
+    {
+        trigger: SEA_EMPEROR_MESSAGE
+    },
+    {
+        trigger: AGARIMOO_MESSAGE
+    },
+    {
+        trigger: CARROT_KING_MESSAGE
+    },
+    {
+        trigger: NURSE_SHARK_MESSAGE
+    },
+    {
+        trigger: BLUE_SHARK_MESSAGE
+    },
+    {
+        trigger: TIGER_SHARK_MESSAGE
+    },
+    {
+        trigger: GREAT_WHITE_SHARK_MESSAGE
+    },
+    {
+        trigger: WEREWOLF_MESSAGE
+    },
+    {
+        trigger: SCARECROW_MESSAGE
+    },
+    {
+        trigger: NIGHTMARE_MESSAGE
+    },
+    {
+        trigger: PHANTOM_FISHER_MESSAGE
+    },
+    {
+        trigger: GRIM_REAPER_MESSAGE
+    },
+    {
+        trigger: FROG_MAN_MESSAGE,
+    },
+    {
+        trigger: TRASH_GOBBLER_MESSAGE,
+    },
+    {
+        trigger: DUMPSTER_DIVER_MESSAGE,
+    },
+    {
+        trigger: BANSHEE_MESSAGE,
+    },
+    {
+        trigger: SNAPPING_TURTLE_MESSAGE,
+    },
+    {
+        trigger: BAYOU_SLUDGE_MESSAGE,
+    },
+    {
+        trigger: ALLIGATOR_MESSAGE,
+    },
+    {
+        trigger: BLUE_RINGED_OCTOPUS_MESSAGE,
     },
 ];