+++ data/data.js
@@ -4,6 +4,24 @@ export const persistentData = new PogObject("FeeshNotifier", {
     "rareCatches": {},
     "totalRareCatches": 0,
     "crimsonIsle": {
+        "fieryScuttler": {
+            "catchesSinceLast": 0,
+            "lastCatchTime": null,
+            "catchesHistory": [],
+            "averageCatches": 0
+        },
+        "ragnarok": {
+            "catchesSinceLast": 0,
+            "lastCatchTime": null,
+            "catchesHistory": [],
+            "averageCatches": 0
+        },
+        "plhlegblast": {
+            "catchesSinceLast": 0,
+            "lastCatchTime": null,
+            "catchesHistory": [],
+            "averageCatches": 0
+        },
         "thunder": {
             "catchesSinceLast": 0,
             "lastCatchTime": null,
@@ -43,6 +61,30 @@ export const persistentData = new PogObject("FeeshNotifier", {
                 "count": 0
             },
         }
+    },
+    "waterHotspotsAndBayou": {
+        "titanoboa": {
+            "catchesSinceLast": 0,
+            "lastCatchTime": null,
+            "catchesHistory": [],
+            "averageCatches": 0
+        },
+        "titanoboaSheds": {
+            "count": 0,
+            "catchesSinceLast": 0,
+            "dropsHistory": []
+        },
+        "wikiTiki": {
+            "catchesSinceLast": 0,
+            "lastCatchTime": null,
+            "catchesHistory": [],
+            "averageCatches": 0
+        },
+        "tikiMasks": {
+            "count": 0,
+            "catchesSinceLast": 0,
+            "dropsHistory": []
+        },
     },
     "fishingProfit": {
         "profitTrackerItems": {},